+++ core/genrcfun.c
@@ -517,7 +517,7 @@ globle long FindMethodByIndex(
 globle void PrintMethod(
   void *theEnv,
   char *buf,
-  int buflen,
+  size_t buflen,
   DEFMETHOD *meth)
   {
 #if MAC_XCD
